+++ Assassin/TransferTask.cpp
@@ -35,18 +35,18 @@ TransferTask::TransferTask(
 			this->m_SaveFile = M2AsyncWait(
 				this->m_SaveFolder->GetFileAsync(this->m_FileName));
 		}
-		this->m_LastStatus = static_cast<TransferTaskStatus>(
+		this->m_Status = static_cast<TransferTaskStatus>(
 			dynamic_cast<Windows::Foundation::IPropertyValue^>(
 				TaskConfig->Lookup(L"LastStatus"))->GetUInt8());
 	}
 	catch (...)
 	{
-		switch (this->m_LastStatus)
+		switch (this->m_Status)
 		{
 		case TransferTaskStatus::Paused:
 		case TransferTaskStatus::Queued:
 		case TransferTaskStatus::Running:
-			this->m_LastStatus = TransferTaskStatus::Error;
+			this->m_Status = TransferTaskStatus::Error;
 			break;
 		default:
 			break;
@@ -68,12 +68,12 @@ TransferTask::TransferTask(
 	}
 	else
 	{
-		switch (this->m_LastStatus)
+		switch (this->m_Status)
 		{
 		case TransferTaskStatus::Paused:
 		case TransferTaskStatus::Queued:
 		case TransferTaskStatus::Running:
-			this->m_LastStatus = TransferTaskStatus::Error;
+			this->m_Status = TransferTaskStatus::Error;
 			break;
 		default:
 			break;
@@ -87,6 +87,75 @@ void TransferTask::RaisePropertyChanged(
 	using Windows::UI::Xaml::Data::PropertyChangedEventArgs;
 	this->PropertyChanged(
 		this, ref new PropertyChangedEventArgs(PropertyName));
+}
+
+void TransferTask::NotifyPropertyChanged()
+{
+	using Windows::Networking::BackgroundTransfer::BackgroundDownloadProgress;
+	
+	if (nullptr != this->m_Operation)
+	{
+		switch (this->m_Operation->Progress.Status)
+		{
+		case BackgroundTransferStatus::Idle:
+			this->m_Status = TransferTaskStatus::Queued;
+			break;
+		case BackgroundTransferStatus::Running:
+		case BackgroundTransferStatus::PausedCostedNetwork:
+		case BackgroundTransferStatus::PausedNoNetwork:
+		case BackgroundTransferStatus::PausedSystemPolicy:
+			this->m_Status = TransferTaskStatus::Running;
+			break;
+		case BackgroundTransferStatus::PausedByApplication:
+			this->m_Status = TransferTaskStatus::Paused;
+			break;
+		case BackgroundTransferStatus::Completed:
+			this->m_Status = TransferTaskStatus::Completed;
+			break;
+		case BackgroundTransferStatus::Canceled:
+			this->m_Status = TransferTaskStatus::Canceled;
+			break;
+		case BackgroundTransferStatus::Error:
+		default:
+			this->m_Status = TransferTaskStatus::Error;
+			break;
+		}
+		this->RaisePropertyChanged(L"Status");
+		if (TransferTaskStatus::Running != this->m_Status)
+			return;
+
+		BackgroundDownloadProgress Progress = this->m_Operation->Progress;
+
+		ULONGLONG LastTickCount = this->m_TickCount;
+		this->m_TickCount = M2GetTickCount();
+
+		uint64 LastBytesReceived = this->m_BytesReceived;
+		this->m_BytesReceived = Progress.BytesReceived;
+		this->RaisePropertyChanged(L"BytesReceived");
+
+		this->m_TotalBytesToReceive = Progress.TotalBytesToReceive;
+		this->RaisePropertyChanged(L"TotalBytesToReceive");
+
+		uint64 DeltaBytesReceived = this->m_BytesReceived - LastBytesReceived;
+		ULONGLONG DeltaPassedTime = this->m_TickCount - LastTickCount;
+		this->m_BytesReceivedSpeed =
+			DeltaBytesReceived * 1000 / DeltaPassedTime;
+		this->RaisePropertyChanged(L"BytesReceivedSpeed");
+
+		if (0 == this->m_BytesReceivedSpeed || 
+			0 == this->m_TotalBytesToReceive)
+		{
+			this->m_RemainTime = static_cast<uint64>(-1);
+		}
+		else
+		{
+			uint64 RemainBytesToReceive = 
+				this->m_TotalBytesToReceive - this->m_BytesReceived;		
+			this->m_RemainTime = 
+				RemainBytesToReceive / this->m_BytesReceivedSpeed;
+		}
+		this->RaisePropertyChanged(L"RemainTime");
+	}
 }
 
 ApplicationDataCompositeValue^ TransferTask::GetTaskConfig()
@@ -132,36 +201,7 @@ IStorageFolder^ TransferTask::SaveFolder::get()
 // The current status of the task.
 TransferTaskStatus TransferTask::Status::get()
 {	
-	if (nullptr != this->m_Operation)
-	{
-		switch (this->m_Operation->Progress.Status)
-		{
-		case BackgroundTransferStatus::Idle:
-			this->m_LastStatus = TransferTaskStatus::Queued;
-			break;
-		case BackgroundTransferStatus::Running:
-		case BackgroundTransferStatus::PausedCostedNetwork:
-		case BackgroundTransferStatus::PausedNoNetwork:
-		case BackgroundTransferStatus::PausedSystemPolicy:
-			this->m_LastStatus = TransferTaskStatus::Running;
-			break;
-		case BackgroundTransferStatus::PausedByApplication:
-			this->m_LastStatus = TransferTaskStatus::Paused;
-			break;
-		case BackgroundTransferStatus::Completed:
-			this->m_LastStatus = TransferTaskStatus::Completed;
-			break;
-		case BackgroundTransferStatus::Canceled:
-			this->m_LastStatus = TransferTaskStatus::Canceled;
-			break;
-		case BackgroundTransferStatus::Error:
-		default:
-			this->m_LastStatus = TransferTaskStatus::Error;
-			break;
-		}
-	}
-
-	return this->m_LastStatus;
+	return this->m_Status;
 }
 
 // The total number of bytes received. This value does not include 
@@ -169,51 +209,26 @@ TransferTaskStatus TransferTask::Status::get()
 // this value may be smaller than in the previous progress report.
 uint64 TransferTask::BytesReceived::get()
 {
-	if (nullptr == this->m_Operation)
-		return 0;
-
-	return this->m_Operation->Progress.BytesReceived;
+	return this->m_BytesReceived;
 }
 
 // The speed of bytes received in one second.
 uint64 TransferTask::BytesReceivedSpeed::get()
 {	
-	if (0 != this->m_LastUpdated)
-	{	
-		uint64 DeltaBytesReceived =
-			this->BytesReceived - this->m_LastBytesReceived;
-		ULONGLONG DeltaPassedTime = 
-			M2GetTickCount() - this->m_LastUpdated;
-
-		this->m_BytesReceivedSpeed = 
-			DeltaBytesReceived * 1000 / DeltaPassedTime;
-	}
-	
-	this->m_LastUpdated = M2GetTickCount();
-	this->m_LastBytesReceived = this->BytesReceived;
-
 	return this->m_BytesReceivedSpeed;
 }
 
 // The remain time, in seconds.
 uint64 TransferTask::RemainTime::get()
 {
-	uint64 ReceivedSpeed = this->m_BytesReceivedSpeed;
-	
-	if (0 == ReceivedSpeed || 0 == this->TotalBytesToReceive)
-		return static_cast<uint64>(-1);
-
-	return (this->TotalBytesToReceive - this->BytesReceived) / ReceivedSpeed;
+	return this->m_RemainTime;
 }
 
 // The total number of bytes of data to download. If this number is
 // unknown, this value is set to 0.
 uint64 TransferTask::TotalBytesToReceive::get()
 {
-	if (nullptr == this->m_Operation)
-		return 0;
-
-	return this->m_Operation->Progress.TotalBytesToReceive;
+	return this->m_TotalBytesToReceive;
 }
 
 // Pauses a download operation.
@@ -231,7 +246,7 @@ void TransferTask::Pause()
 		}
 		else
 		{
-			this->m_LastStatus = TransferTaskStatus::Error;
+			this->m_Status = TransferTaskStatus::Error;
 		}
 	}
 }
@@ -252,7 +267,7 @@ void TransferTask::Resume()
 		}
 		else
 		{
-			this->m_LastStatus = TransferTaskStatus::Error;
+			this->m_Status = TransferTaskStatus::Error;
 		}
 	}
 }
@@ -275,7 +290,7 @@ void TransferTask::Cancel()
 		}
 		else
 		{
-			this->m_LastStatus = TransferTaskStatus::Error;
+			this->m_Status = TransferTaskStatus::Error;
 		}
 		break;
 	default: