+++ Nagisa/MainPage.xaml.h
@@ -10,24 +10,38 @@
 
 namespace Nagisa
 {
-	using Assassin::TransferManager;
+	using Assassin::ITransferManager;
 	using Platform::Object;
 	using Platform::String;
+	using Windows::Foundation::IAsyncOperation;
 	using Windows::UI::Xaml::RoutedEventArgs;
 	using Windows::UI::Xaml::Controls::AutoSuggestBox;
 	using Windows::UI::Xaml::Controls::AutoSuggestBoxQuerySubmittedEventArgs;
 	using Windows::UI::Xaml::Controls::AutoSuggestBoxTextChangedEventArgs;
+	using Windows::UI::Xaml::Controls::ContentDialog;
+	using Windows::UI::Xaml::Controls::ContentDialogResult;
+	using Windows::UI::Xaml::Data::INotifyPropertyChanged;
+	using Windows::UI::Xaml::Data::PropertyChangedEventHandler;
 
 	public ref class MainPage sealed
 	{
 	public:
 		MainPage();
-	internal:
-		TransferManager ^ m_TransferManager = nullptr;
+
+		property ITransferManager^ Manager
+		{
+			ITransferManager^ get();
+		}
+
 	private:
+		ITransferManager ^ m_TransferManager = nullptr;
+
 		void RefreshTaskList();
+		void RefreshTaskListAsync();
 		void SearchTaskList(
 			String^ SearchFilter);
+		IAsyncOperation<ContentDialogResult>^ ShowContentDialogAsync(
+			ContentDialog^ Dialog);
 
 		void AboutButton_Click(
 			Object^ sender,