+++ Nagisa/SettingsDialog.xaml.h
@@ -9,8 +9,9 @@
 
 namespace Nagisa
 {
-	using Assassin::TransferManager;
+	using Assassin::ITransferManager;
 	using Platform::Object;
+	using Windows::Storage::IStorageFolder;
 	using Windows::UI::Xaml::Controls::ContentDialog;
 	using Windows::UI::Xaml::Controls::ContentDialogButtonClickEventArgs;
 	using Windows::UI::Xaml::RoutedEventArgs;
@@ -19,13 +20,21 @@ namespace Nagisa
 	public ref class SettingsDialog sealed
 	{
 	public:
-		SettingsDialog();
-	internal:
-		TransferManager^ m_TransferManager = nullptr;
+		SettingsDialog(
+			ITransferManager^ TransferManager);
 	private:
+		ITransferManager ^ m_TransferManager = nullptr;
+
+		IStorageFolder^ m_LastusedFolder = nullptr;
 
 		void ContentDialog_Loaded(
 			Object^ sender, 
+			RoutedEventArgs^ e);
+		void CustomDownloadFolderBrowseButton_Click(
+			Object^ sender,
+			RoutedEventArgs^ e);
+		void UseCustomFolder_Click(
+			Object^ sender,
 			RoutedEventArgs^ e);
 	};
 }